rename logEvent to formatEventDataAndSave refacot into separate file
