+++ lib/dataUtils.js
@@ -12,6 +12,10 @@ var logMessage = function(debug, functionName, message, jsonObject) {
   }
 };
 
+var timeTookInSeconds = function (startTime, endTime) {
+  return ( (endTime - startTime) / 1000.0 ) + " seconds"
+}
+
 function _hashSensitive(jsonBody, debug) {
   if (jsonBody === null) return jsonBody;
 
@@ -214,6 +218,20 @@ function _getEventModelFromRequestandResponse(
   return logData;
 }
 
+function isJsonHeader(msg) {
+  if (msg) {
+    var headers = msg.headers || msg._moHeaders;
+    if (headers['content-encoding']) {
+      return false;
+    }
+    if (headers['content-type'] && headers['content-type'].indexOf('json') >= 0) {
+      return true;
+    }
+  }
+  return false;
+}
+
+
 module.exports = {
   getUrlFromRequestOptions: _getUrlFromRequestOptions,
   getEventModelFromRequestandResponse: _getEventModelFromRequestandResponse,
@@ -221,5 +239,7 @@ module.exports = {
   startWithJson: _startWithJson,
   bodyToBase64: _bodyToBase64,
   hashSensitive: _hashSensitive,
-  logMessage: logMessage
+  logMessage: logMessage,
+  timeTookInSeconds: timeTookInSeconds,
+  isJsonHeader: isJsonHeader
 };
