+++ lib/index.js
@@ -21,23 +21,56 @@ var moesifConfigManager = require('./moesifConfigManager');
 var uuid4 = require('uuid4');
 var unparsed = require('koa-body/unparsed.js');
 var pjson = require('../package.json');
+var ensureValidUtils = require('./ensureValidUtils');
+var formatEventDataAndSave = require('./formatEventDataAndSave');
 
 // express converts headers to lowercase
 const TRANSACTION_ID_HEADER = 'x-moesif-transaction-id';
 
-
-exports.defaultSkip = function(req, res) {
-  return false;
-};
-
-//
-// ### function moesif(options)
-// #### @options {Object} options to initialize the middleware.
-//
-
 var logMessage = dataUtils.logMessage;
+var timeTookInSeconds = dataUtils.timeTookInSeconds;
+var safeJsonParse = dataUtils.safeJsonParse;
+var bodyToBase64 = dataUtils.bodyToBase64;
+var startWithJson = dataUtils.startWithJson;
+var isJsonHeader = dataUtils.isJsonHeader;
+
+var ensureValidOptions = ensureValidUtils.ensureValidOptions;
+var ensureValidUserModel = ensureValidUtils.ensureValidUserModel;
+var ensureValidUsersBatchModel = ensureValidUtils.ensureValidUsersBatchModel;
+var ensureValidCompanyModel = ensureValidUtils.ensureValidCompanyModel;
+var ensureValidCompaniesBatchModel = ensureValidUtils.ensureValidCompaniesBatchModel;
+
+
+function appendChunk(buf, chunk) {
+  if (chunk) {
+    var appendChunkFunctionStartTime = Date.now();
+    if (Buffer.isBuffer(chunk)) {
+      return buf ? Buffer.concat([buf, chunk]) : Buffer.from(chunk);
+    } else if (typeof chunk === 'string') {
+      return buf ? Buffer.concat([buf, Buffer.from(chunk)]) : Buffer.from(chunk);
+    } else if (typeof chunk === 'object' || Array.isArray(chunk)) {
+      try {
+        return buf
+          ? Buffer.concat([buf, Buffer.from(JSON.stringify(chunk))])
+          : Buffer.from(JSON.stringify(chunk));
+      } catch (err) {
+        return buf;
+      }
+    } else {
+      console.error('Response body chunk is not a Buffer or String.');
+    }
+    var appendChunkFunctionEndTime = Date.now();
+    logMessage(options.debug, 'appendChunkFunction took time ', timeTookInSeconds(appendChunkFunctionStartTime, appendChunkFunctionEndTime));
+  }
+}
+
+// default option utility functions.
 
 var noop = function () {}; // implicitly return undefined
+
+var defaultSkip = function(req, res) {
+  return false;
+};
 
 var defaultIdentifyUser = function (req, res) {
   if (req) {
@@ -53,31 +86,10 @@ var defaultIdentifyUser = function (req, res) {
   return undefined;
 };
 
-function ensureToString(id) {
-  if (typeof id === 'number') {
-    return String(id);
-  }
-  if (typeof id === 'string') {
-    return id;
-  }
-  if (id === null || id === undefined) {
-    return id;
-  }
-  if (typeof id === 'object') {
-    return String(id);
-  }
-  return id;
-}
-
-function timeTookInSeconds(startTime, endTime) {
-  return ( (endTime - startTime) / 1000.0 ) + " seconds"
-}
-
-var hashSensitive = dataUtils.hashSensitive;
-var safeJsonParse = dataUtils.safeJsonParse;
-var bodyToBase64 = dataUtils.bodyToBase64;
-var startWithJson = dataUtils.startWithJson;
-
+//
+// ### function moesif(options)
+// #### @options {Object} options to initialize the middleware.
+//
 module.exports = function(options) {
   logMessage(options.debug, 'moesifInitiator', 'start');
 
@@ -387,7 +399,7 @@ module.exports = function(options) {
             })
             .then(() => {
               var logEventAfterGettingRawBodyStartTime = Date.now();
-              logEventOnResponseEnd(finalBuf, req, res, options, trySaveEventLocal);
+              formatEventDataAndSave(finalBuf, req, res, options, trySaveEventLocal);
               var logEventAfterGettingRawBodyEndTime = Date.now();
               logMessage(options.debug, 'logEventAfterGettingRawBody took time ', timeTookInSeconds(logEventAfterGettingRawBodyStartTime, logEventAfterGettingRawBodyEndTime));
             })
@@ -396,7 +408,7 @@ module.exports = function(options) {
             });
         } else {
           var logEventWithoutGettingRawBodyStartTime = Date.now();
-          logEventOnResponseEnd(finalBuf, req, res, options, trySaveEventLocal);
+          formatEventDataAndSave(finalBuf, req, res, options, trySaveEventLocal);
           var logEventWithoutGettingRawBodyEndTime = Date.now();
           logMessage(options.debug, 'logEventWithoutGettingRawBody took time ', timeTookInSeconds(logEventWithoutGettingRawBodyStartTime, logEventWithoutGettingRawBodyEndTime));
         }
@@ -470,337 +482,4 @@ module.exports = function(options) {
   return moesifMiddleware;
 };
 
-// START Helper functions.
 
-function appendChunk(buf, chunk) {
-  if (chunk) {
-    var appendChunkFunctionStartTime = Date.now();
-    if (Buffer.isBuffer(chunk)) {
-      return buf ? Buffer.concat([buf, chunk]) : Buffer.from(chunk);
-    } else if (typeof chunk === 'string') {
-      return buf ? Buffer.concat([buf, Buffer.from(chunk)]) : Buffer.from(chunk);
-    } else if (typeof chunk === 'object' || Array.isArray(chunk)) {
-      try {
-        return buf
-          ? Buffer.concat([buf, Buffer.from(JSON.stringify(chunk))])
-          : Buffer.from(JSON.stringify(chunk));
-      } catch (err) {
-        return buf;
-      }
-    } else {
-      console.error('Response body chunk is not a Buffer or String.');
-    }
-    var appendChunkFunctionEndTime = Date.now();
-    logMessage(options.debug, 'appendChunkFunction took time ', timeTookInSeconds(appendChunkFunctionStartTime, appendChunkFunctionEndTime));
-  }
-}
-
-function safeGetResponseHeaders(res) {
-  try {
-    if (res.getHeaders) {
-      return res.getHeaders();
-    }
-    try {
-      // access ._headers will result in exception
-      // in some versions fo node.
-      // so must be in try block.
-      if (res._headers) {
-        return res._headers;
-      }
-    } catch (err) {
-    }
-    return res.headers || decodeHeaders(res._header);
-  } catch(err) {
-    return {};
-  }
-}
-
-function logEventOnResponseEnd(chunk, req, res, options, saveEvent) {
-  logMessage(options.debug, 'logEventOnResponseEnd', 'reqUrl=' + req.originalUrl);
-  logMessage(options.debug, 'logEventOnResponseEnd', 'chunk=', chunk);
-  // res.time = new Date();
-  // res.responseTime = new Date() - req._startTime;
-
-  var logData = {};
-  logData.request = {};
-  logData.request.verb = req.method;
-  var protocol =
-    (req.connection && req.connection.encrypted) || req.secure ? 'https://' : 'http://';
-
-  var host = req.headers.host || req.hostname;
-  logData.request.uri = protocol + host + (req.originalUrl || req.url);
-  logData.request.headers = req.headers;
-
-  if (options.logBody) {
-    var parseRequestBodyStartTime = Date.now();
-    if (req._moBody) {
-      var parseRequestBodyAsTransferEncodingStartTime = Date.now();
-      logMessage(options.debug, 'logEventOnResponseEnd', 'processing req._moBody');
-      logData.request.transferEncoding = req._moTransferEncoding;
-      logData.request.body = req._moBody;
-      var parseRequestBodyAsTransferEncodingEndTime = Date.now();
-      logMessage(options.debug, 'parseRequestBodyAsTransferEncoding took time ', timeTookInSeconds(parseRequestBodyAsTransferEncodingStartTime, parseRequestBodyAsTransferEncodingEndTime));
-    } else if (req.body && (Number(Buffer.byteLength(JSON.stringify(req.body)) < options.requestMaxBodySize))) {
-      logMessage(options.debug, 'logEventOnResponseEnd', 'processing req.body');
-      var isReqBodyMaybeJson = isJsonHeader(req) || startWithJson(req.body);
-
-      if (isReqBodyMaybeJson) {
-        var parseRequestBodyAsJsonStartTime = Date.now();
-        var parsedReqBody = safeJsonParse(req.body);
-
-        logData.request.transferEncoding = parsedReqBody.transferEncoding;
-        logData.request.body = parsedReqBody.body;
-        var parseRequestBodyAsJsonEndTime = Date.now();
-        logMessage(options.debug, 'parseRequestBodyAsJson took time ', timeTookInSeconds(parseRequestBodyAsJsonStartTime, parseRequestBodyAsJsonEndTime));
-      } else {
-        var parseRequestBodyAsBase64StartTime = Date.now();
-        logData.request.transferEncoding = 'base64';
-        logData.request.body = bodyToBase64(req.body);
-        var parseRequestBodyAsBase64EndTime = Date.now();
-        logMessage(options.debug, 'parseRequestBodyAsBase64 took time ', timeTookInSeconds(parseRequestBodyAsBase64StartTime, parseRequestBodyAsBase64EndTime));
-      }
-    }
-    var parseRequestBodyEndTime = Date.now();
-    logMessage(options.debug, 'parseRequestBody took time ', timeTookInSeconds(parseRequestBodyStartTime, parseRequestBodyEndTime));
-  }
-
-  logData.request.ipAddress = requestIp.getClientIp(req);
-  logData.request.apiVersion = options.apiVersion;
-  logData.request.time = req._startTime;
-
-  logData.response = {};
-  logData.response.status = res.statusCode ? res.statusCode : 599;
-  res._moHeaders = safeGetResponseHeaders(res);
-  logData.response.headers = res._moHeaders;
-  logData.response.time = res.time;
-
-  if (options.logBody) {
-    if (chunk) {
-      logMessage(options.debug, 'logEvent', 'processing chunk');
-      if (chunk.length < options.responseMaxBodySize) {
-        if (isJsonHeader(res) || startWithJson(chunk)) {
-          var parsedResBody = safeJsonParse(chunk);
-          logData.response.transferEncoding = parsedResBody.transferEncoding;
-          logData.response.body = parsedResBody.body;
-        } else {
-          logData.response.transferEncoding = 'base64';
-          logData.response.body = bodyToBase64(chunk);
-        }
-      } else {
-        logData.response.body = {
-          msg: 'response.body.length is over options responseMaxBodySize of ' + options.responseMaxBodySize
-        }
-      }
-    }
-  }
-
-  logMessage(options.debug, 'logEvent', 'created data=', logData);
-
-  logData = options.maskContent(logData);
-
-  var identifyUserStartTime = Date.now();
-  logData.userId = ensureToString(options.identifyUser(req, res));
-  var identifyUserEndTime = Date.now();
-  logMessage(options.debug, 'identifyUser took time ', timeTookInSeconds(identifyUserStartTime, identifyUserEndTime));
-
-  var identifyCompanyStartTime = Date.now();
-  logData.companyId = ensureToString(options.identifyCompany(req, res));
-  var identifyCompanyEndTime = Date.now();
-  logMessage(options.debug, 'identifyCompany took time ', timeTookInSeconds(identifyCompanyStartTime, identifyCompanyEndTime));
-
-  logData.sessionToken = options.getSessionToken(req, res);
-  logData.tags = options.getTags(req, res);
-  logData.request.apiVersion = options.getApiVersion(req, res);
-  logData.metadata = options.getMetadata(req, res);
-
-  // Set API direction
-  logData.direction = "Incoming"
-
-  logMessage(options.debug, 'logEvent', 'applied options to data=', logData);
-
-  var ensureValidLogDataStartTime = Date.now();
-  ensureValidLogData(logData);
-  var ensureValidLogDataEndTime = Date.now();
-  logMessage(options.debug, 'ensureValidLogData took time ', timeTookInSeconds(ensureValidLogDataStartTime, ensureValidLogDataEndTime));
-
-  // This is fire and forget, we don't want logging to hold up the request so don't wait for the callback
-  if (!options.skip(req, res)) {
-    logMessage(options.debug, 'logEvent', 'queue data to send to moesif');
-
-    if (!options.noAutoHideSensitive) {
-      var noAutoHideSensitiveStartTime = Date.now();
-      // autoHide
-      try {
-        logData.request.headers = hashSensitive(logData.request.headers, options.debug);
-        logData.request.body = hashSensitive(logData.request.body, options.debug);
-        logData.response.headers = hashSensitive(logData.response.headers, options.debug);
-        logData.response.body = hashSensitive(logData.response.body, options.debug);
-      } catch (err) {
-        logMessage(options.debug, 'logEvent', 'error on hashSensitive err=' + err);
-      }
-      var noAutoHideSensitiveEndTime = Date.now();
-      logMessage(options.debug, 'noAutoHideSensitive took time ', timeTookInSeconds(noAutoHideSensitiveStartTime, noAutoHideSensitiveEndTime));
-    }
-
-    // Add Transaction Id to Event Request Model
-    if (logData.response.headers[TRANSACTION_ID_HEADER]) {
-      logData.request.headers[TRANSACTION_ID_HEADER] = logData.response.headers[TRANSACTION_ID_HEADER];
-    }
-
-    saveEvent(logData);
-  }
-}
-
-function isJsonHeader(msg) {
-  if (msg) {
-    var headers = msg.headers || msg._moHeaders;
-    if (headers['content-encoding']) {
-      return false;
-    }
-    if (headers['content-type'] && headers['content-type'].indexOf('json') >= 0) {
-      return true;
-    }
-  }
-  return false;
-}
-
-function decodeHeaders(header) {
-  try {
-    var keyVal = header.split('\r\n');
-
-    // Remove Request Line or Status Line
-    keyVal.shift();
-
-    var obj = {};
-    var i;
-    for (i in keyVal) {
-      keyVal[i] = keyVal[i].split(':', 2);
-      if (keyVal[i].length != 2) {
-        continue;
-      }
-      obj[keyVal[i][0].trim()] = keyVal[i][1].trim();
-    }
-    return obj;
-  } catch (err) {
-    return {};
-  }
-}
-
-function ensureValidOptions(options) {
-  if (!options) throw new Error('options are required by moesif-nodejs middleware');
-  if (!options.applicationId || typeof options.applicationId !== 'string') {
-    throw new Error(
-      'A moesif application id is required. Please obtain it through your settings at www.moesif.com'
-    );
-  }
-  if (options.applicationId.length < 50) {
-    throw new Error(
-      'A moesif application id is required. The format of the moesif application id provided does not look correct. Please obtain it through your settings at www.moesif.com'
-    );
-  }
-  if (options.identifyUser && !isFunction(options.identifyUser)) {
-    throw new Error('identifyUser should be a function');
-  }
-  if (options.identifyCompany && !isFunction(options.identifyCompany)) {
-    throw new Error('identifyCompany should be a function');
-  }
-  if (options.getMetadata && !isFunction(options.getMetadata)) {
-    throw new Error('getMetadata should be a function');
-  }
-  if (options.getSessionToken && !isFunction(options.getSessionToken)) {
-    throw new Error('getSessionToken should be a function');
-  }
-  if (options.getTags && !isFunction(options.getTags)) {
-    throw new Error('getTags should be a function');
-  }
-  if (options.getApiVersion && !isFunction(options.getApiVersion)) {
-    throw new Error('getApiVersion should be a function');
-  }
-  if (options.maskContent && !isFunction(options.maskContent)) {
-    throw new Error('maskContent should be a function');
-  }
-  if (options.skip && !isFunction(options.skip)) {
-    throw new Error('skip should be a function');
-  }
-  if (options.retry && (!isNumber(options.retry) || options.retry > 3 || options.retry < 0)) {
-    throw new Error('If retry is set, it must be a number between 0 to 3.');
-  }
-  if (options.batchSize && (!isNumber(options.batchSize) || options.batchSize <= 1)) {
-    throw new Error('batchSize must be a number greater than or equal to 1');
-  }
-  if (options.batchMaxTime && (!isNumber(options.batchMaxTime) || options.batchMaxTime <= 500)) {
-    throw new Error('batchMaxTime must be greater than 500 milliseonds');
-  }
-  if (options.requestMaxBodySize && (!isNumber(options.requestMaxBodySize) || options.requestMaxBodySize < 0)) {
-    throw new Error('requestMaxBodySize must be a number greater than 0');
-  }
-  if (options.responseMaxBodySize && (!isNumber(options.responseMaxBodySize) || options.responseMaxBodySize < 0)) {
-    throw new Error('responseMaxBodySize must be a number greater than 0');
-  }
-}
-
-function ensureValidLogData(logData) {
-  if (!logData.request) {
-    throw new Error(
-      'For Moesif events, request and response objects are required. Please check your maskContent function do not remove this'
-    );
-  } else {
-    if (!logData.request.time) {
-      throw new Error(
-        'For Moesif events, request time is required. Please check your maskContent function do not remove this'
-      );
-    }
-    if (!logData.request.verb) {
-      throw new Error(
-        'For Moesif events, request verb is required. Please check your maskContent function do not remove this'
-      );
-    }
-    if (!logData.request.uri) {
-      throw new Error(
-        'For Moesif events, request uri is required. Please check your maskContent function do not remove this'
-      );
-    }
-  }
-  if (!logData.response) {
-    throw new Error(
-      'For Moesif events, request and response objects are required. Please check your maskContent function do not remove this'
-    );
-  } else {
-    // if (!logData.response.body) {
-    //   throw new Error('for log events, response body objects is required but can be empty object');
-    // }
-    if (!logData.request.time) {
-      throw new Error(
-        'For Moesif events, response time is required. The middleware should populate it automatically. Please check your maskContent function do not remove this'
-      );
-    }
-  }
-}
-
-function ensureValidUserModel(userModel) {
-  if (!userModel.userId) {
-    throw new Error('To update a user, a userId field is required');
-  }
-}
-
-function ensureValidUsersBatchModel(usersBatchModel) {
-  for (let userModel of usersBatchModel) {
-    if (!userModel.userId) {
-      throw new Error('To update a user, a userId field is required');
-    }
-  }
-}
-
-function ensureValidCompanyModel(companyModel) {
-  if (!companyModel.companyId) {
-    throw new Error('To update a company, a companyId field is required');
-  }
-}
-
-function ensureValidCompaniesBatchModel(companiesBatchModel) {
-  for (let companyModel of companiesBatchModel) {
-    if (!companyModel.companyId) {
-      throw new Error('To update a company, a companyId field is required');
-    }
-  }
-}
